implement isyntax support under windows

+++ src/b64.c
@@ -177,7 +177,7 @@ unsigned char *b64_decode_ex(const char *src, size_t len, size_t *decsize) {
 }
 
 unsigned char *b64_encode(const unsigned char *src, size_t len) {
-    int32_t byte_length = 0;
+    int8_t byte_length = 0;
     unsigned char *encoded_buffer = NULL;
     int8_t num_of_chars_to_enc = 3;
     int8_t sections = 4;
